Update eggScr.py
target website changed the CSS selector for the location. I am looking into writing functionality so that prior to scraping the prices, the program first compares the current HTML content with the last version and if there are any changes it will let you know and reccomend the change to make so that the code works.

Realistically I would like to find a better way of scraping the results as under the current system, the program is susceptible to changes in website structure/code. For now I will try the above solution and see if that works for now. Although I admit it feels like a roundabout solution to the issue.